initial cogification
