+++ code/constants.py
@@ -1,0 +1,7 @@
+PLAIN_DAYS = ['ma', 'ti', 'ke', 'to', 'pe', 'la', 'su']
+
+#Emojis
+CHECK_MARK = "✅"
+DENY_EMOJI = "👎"
+
+token = "app.token"
