+++ code/doodle/constants.py
@@ -1,0 +1,2 @@
+DOODLE_NEW_MESSAGE = "<@& 543751882104111112> <@& 579384808996733122> doodle"
+DOODLE_INVALID_SUBCOMMAND = "Invalid doodle command. See help."
