+++ code/prac/constants.py
@@ -1,0 +1,7 @@
+PRAC_ADD_OK = "New prac session created."
+PRAC_SUB_OK = "Player substitution successfull."
+PRAC_MOVE_OK = "Session time change successfull."
+PRAC_MISSING_TIME = "Missing mandatory time argument."
+PRAC_INVALID_SUBCOMMAND = "Invalid prac command. See help."
+
+database = "pracs.db"
