--- constants.py
@@ -1,21 +1,0 @@
-PLAIN_DAYS = ['ma', 'ti', 'ke', 'to', 'pe', 'la', 'su']
-
-#Emojis
-CHECK_MARK = "✅"
-DENY_EMOJI = "👎"
-
-# Messages
-## Prac
-PRAC_ADD_OK = "New prac session created."
-PRAC_SUB_OK = "Player substitution successfull."
-PRAC_MOVE_OK = "Session time change successfull."
-PRAC_MISSING_TIME = "Missing mandatory time argument."
-PRAC_INVALID_SUBCOMMAND = "Invalid prac command. See help."
-
-## Doodle
-DOODLE_NEW_MESSAGE = "<@&543751882104111112> <@&579384808996733122> doodle"
-DOODLE_INVALID_SUBCOMMAND = "Invalid doodle command. See help."
-
-# File names
-token = "app.token"
-database = "pracs.db"
